Rearrange the reverse_words code to make it easier to resume
